Create Pry::Platform and remove some methods.
Remove these methods from Pry::Helpers::BaseHelpers to Pry::Platform
- windows?
- windows_ansi?
- mri?
- mri_20?

Other methods will remove to Pry::Platform later, in this commit remove
methods which are needed to only Pry.load_win32console.
