+++ lib/pry.rb
@@ -5,6 +5,7 @@ require 'pp'
 
 require 'pry/input_lock'
 require 'pry/exceptions'
+require 'pry/platform'
 require 'pry/helpers/base_helpers'
 require 'pry/hooks'
