+++ lib/pry/indent.rb
@@ -392,7 +392,7 @@ class Pry
       cols = cols.to_i
       lines = (cols != 0 ? (line_to_measure.length / cols + 1) : 1).to_i
 
-      if Pry::Helpers::BaseHelpers.windows_ansi?
+      if Pry::Platform.windows_ansi?
         move_up   = "\e[#{lines}F"
         move_down = "\e[#{lines}E"
       else
